refactor: relocate function exporting data to files or other entity
